pyproject.toml
[tool.poetry]
name = "enterplatform"
version = "0.1.0"
description = ""
authors = ["lee-gilro <rlfgussh@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
mysqlclient = "^2.1.1"
django-baton = {extras = ["analytics"], version = "^2.6.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
